requirements.txt
# Core Qt Framework
PyQt6>=6.5.0
PyQt6-Qt6>=6.5.0
PyQt6-sip>=13.5.0

# Database
SQLAlchemy>=2.0.0
alembic>=1.12.0

# Audio Processing
pygame>=2.5.0
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1
python-vlc>=3.0.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Audio Metadata
mutagen>=1.46.0
eyed3>=0.9.7

# Image Processing
Pillow>=10.0.0

# Visualization
pyqtgraph>=0.13.0
matplotlib>=3.7.0

# Async Operations
aiofiles>=23.0.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
watchdog>=3.0.0

# Development & Testing
pytest>=7.4.0
pytest-qt>=4.2.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Build & Distribution
pyinstaller>=6.0.0
cx_Freeze>=6.15.0

# Logging
loguru>=0.7.0

# Configuration
pyyaml>=6.0.1
toml>=0.10.2

# AI/ML Integration (Optional but recommended)
openai>=1.0.0  # For advanced metadata enrichment

# Performance
numba>=0.58.0
cython>=3.0.0